Refactor wan   -  Addressed spec file issue   -  Adding compile only flag by default   -  Refactor of common utils
