+++ QEfficient/diffusers/pipelines/configs/wan_config.json
@@ -1,43 +1,46 @@
 {
-    "description": "Default configuration for Wan pipeline",
-    "model_type": "wan",
+    "description": "Default configuration for Wan unified transformer",
     "modules": {
         "transformer": {
             "specializations": [
-                {
-                    "batch_size": "1",
-                    "num_channels": "16",
-                    "num_frames": "21",
-                    "latent_height": "24",
-                    "latent_width": "40",
-                    "steps": "1",
-                    "sequence_length": "512",
-                    "cl": "5040",
-                    "model_type": 1,
-                },
-                {
-                    "batch_size": "1",
-                    "num_channels": "16",
-                    "num_frames": "21",
-                    "latent_height": "24",
-                    "latent_width": "40",
-                    "steps": "1",
-                    "sequence_length": "512",
-                    "cl": "5040",
-                    "model_type": 2,
-                },
-            ],
-            "compilation": {
-                "onnx_path": null,
-                "compile_dir": null,
-                "mdp_ts_num_devices": 16,
-                "mxfp6_matmul": true,
-                "convert_to_fp16": true,
-                "aic_num_cores": 16,
-                "mos": 1,
-                "mdts_mos": 1,
-            },
-            "execute": {"device_ids": null},
+                                {
+                                "batch_size": "1",
+                                "num_channels": "16",
+                                "num_frames": "21",
+                                "latent_height": "24",
+                                "latent_width": "40",
+                                "steps": "1",
+                                "sequence_length": "512",
+                                "cl": "5040",
+                                "model_type": 1
+                                },
+                                {
+                                "batch_size": "1",
+                                "num_channels": "16",
+                                "num_frames": "21",
+                                "latent_height": "24",
+                                "latent_width": "40",
+                                "steps": "1",
+                                "sequence_length": "512",
+                                "cl": "5040",
+                                "model_type": 2
+                                }
+                            ],
+            "compilation":
+                          {
+                            "onnx_path": null,
+                            "compile_dir": null,
+                            "mdp_ts_num_devices": 16,
+                            "mxfp6_matmul": true,
+                            "convert_to_fp16": true,
+                            "aic_num_cores": 16,
+                            "mos": 1,
+                            "mdts_mos": 1
+                            },
+            "execute":
+                      {
+                        "device_ids": null
+                      }
         }
     },
 }
