+++ pyproject.toml
@@ -20,6 +20,7 @@ classifiers = [
 requires-python = ">=3.8,<3.11"
 dependencies = [
     "transformers==4.55.0",
+    "diffusers== 0.35.1",
     "huggingface-hub==0.34.0",
     "hf_transfer==0.1.9",
     "peft==0.17.0",
@@ -39,6 +40,9 @@ dependencies = [
     "fire",
     "py7zr",
     "torchmetrics==1.7.0",
+    "ftfy==6.3.1",
+    "imageio==2.37.2",
+    "imageio-ffmpeg==0.6.0",
     "torch==2.7.0; platform_machine=='aarch64'",
     # Specifying torch cpu package URL per python version, update the list once pytorch releases whl for python>3.11
     "torch@https://download.pytorch.org/whl/cpu/torch-2.4.1%2Bcpu-cp38-cp38-linux_x86_64.whl ; python_version=='3.8' and platform_machine=='x86_64'",
@@ -52,12 +56,6 @@ dependencies = [
 test = ["pytest","pytest-mock"]
 docs = ["Sphinx==7.1.2","sphinx-rtd-theme==2.0.0","myst-parser==3.0.1","sphinx-multiversion"]
 quality = ["black", "ruff", "hf_doc_builder@git+https://github.com/huggingface/doc-builder.git"]
-diffusers = [
-    "diffusers== 0.35.1",
-    "ftfy==6.3.1",
-    "imageio==2.37.2",
-    "imageio-ffmpeg==0.6.0"
-    ]
 [build-system]
 requires = ["setuptools>=62.0.0"]
 build-backend = "setuptools.build_meta"
